+++ src/node/secrets.h
@@ -18,26 +18,35 @@ namespace ccf
     // Encrypted secret for each backup
     std::vector<uint8_t> encrypted_secret = {};
 
-    MSGPACK_DEFINE(version, encrypted_secret);
+    // Version at which the previous secret is stored at
+    std::optional<kv::Version> previous_secret_stored_version = std::nullopt;
+
+    MSGPACK_DEFINE(version, encrypted_secret, previous_secret_stored_version);
   };
 
-  DECLARE_JSON_TYPE(EncryptedLedgerSecret)
+  DECLARE_JSON_TYPE_WITH_OPTIONAL_FIELDS(EncryptedLedgerSecret)
   DECLARE_JSON_REQUIRED_FIELDS(EncryptedLedgerSecret, version, encrypted_secret)
+  DECLARE_JSON_OPTIONAL_FIELDS(
+    EncryptedLedgerSecret, previous_secret_stored_version)
 
-  struct EncryptedLedgerSecrets
+  using EncryptedLedgerSecrets = std::vector<EncryptedLedgerSecret>;
+  using SecretsForNodes = std::map<NodeId, EncryptedLedgerSecrets>;
+
+  struct EncryptedLedgerSecretsNodesInfo
   {
     std::vector<uint8_t> primary_public_encryption_key = {};
 
-    std::vector<EncryptedLedgerSecret> encrypted_secrets = {};
+    SecretsForNodes secrets_for_nodes = {};
 
-    MSGPACK_DEFINE(primary_public_encryption_key, encrypted_secrets);
+    MSGPACK_DEFINE(primary_public_encryption_key, secrets_for_nodes)
   };
-
-  DECLARE_JSON_TYPE(EncryptedLedgerSecrets)
+  DECLARE_JSON_TYPE(EncryptedLedgerSecretsNodesInfo)
   DECLARE_JSON_REQUIRED_FIELDS(
-    EncryptedLedgerSecrets, primary_public_encryption_key, encrypted_secrets)
+    EncryptedLedgerSecretsNodesInfo,
+    primary_public_encryption_key,
+    secrets_for_nodes)
 
   // This map is used to communicate encrypted ledger secrets from the primary
   // to the backups during recovery (past secrets) and re-keying (new secret)
-  using Secrets = kv::Map<NodeId, EncryptedLedgerSecrets>;
+  using Secrets = kv::Map<size_t, EncryptedLedgerSecretsNodesInfo>;
 }